README.hardware: Drop the NVME support
Issue: LIN1021-2342

The new version bootloader has updated the PCIe controller compatible
from "marvell,pci-host-octeontx2-pem" to "pci-host-ecam-generic".
That means that the pci-host-generic driver will be used instead of
the pci-octeontx2-pem. But the pci-host-generic doesn't include the
workaround for the PEM config writes. That is the reason why the NVME
can't work. The fix for the PEM config writes depend on the old version
bootloader, so we don't have an elegant method to workaround this issue.
And since the SDK has abandoned to support the old version SoC for a
very long time and there is also no NVME interface in the new version
board, it is time for us to drop the support of NVME too.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>